登录，前端构建

+++ autumn-gateway-dashboard/src/main/java/xyz/wewin/autumn/gateway/dashboard/service/ApplicationService.java
@@ -32,7 +32,7 @@ public class ApplicationService {
                                   int page,
                                   int size) {
         Pageable pageable = PageRequest.of(page - 1, size);
-        long offset = (page - 1) * size;
+        long offset = (long) (page - 1) * size;
         long limit = size;
         List<Application> content = repository.findWithPage(appId, name, offset, limit);
         long total = repository.countByFilter(appId, name);
