+++ autumn-gateway-dashboard/src/main/java/xyz/wewin/autumn/gateway/dashboard/service/PermissionService.java
@@ -23,7 +23,7 @@ public class PermissionService {
 
     public Page<Permission> list(int page, int size, Long appId, String code, String name, String permType) {
 
-        long offset = (page - 1) * size;
+        long offset = (long) (page - 1) * size;
         long limit = size;
         List<Permission> content = permissionRepository.findWithPage(appId,
                 code,
