+++ autumn-gateway-dashboard/src/main/java/xyz/wewin/autumn/gateway/dashboard/service/RoleService.java
@@ -31,7 +31,7 @@ public class RoleService {
     @Autowired
     private UserRepository userRepository;
     public Page<Role> list(int page, int size, Long appId, String code, String name) {
-        long offset = (page - 1) * size;
+        long offset = (long) (page - 1) * size;
         long limit = size;
         List<Role> content = roleRepository.findWithPage(appId,
                 code,
